+++ protoc_plugin/CHANGELOG.md
@@ -209,7 +209,7 @@
 ## 14.0.1
 
 * Remove the benchmark from the protoc_package. It now lives in
-  [https://github.com/dart-lang/protobuf] as `api_benchmark`.
+  [https://github.com/google/protobuf.dart] as `api_benchmark`.
   This simplifies the dev_dependencies of this package.
 
 ## 14.0.0